Update of dependencies
Really the whole bot needs a rewrite. The requests library has been
discontinued since 2020, and there is real value to adding this
as part of a proper bot with interactions and all that jazz.

But at the very least, it's just a bit newer now, and definitely
is using a newer tumblr.js library which should help stave off
obsolesence a few more months.